+++ .gitignore
@@ -37,5 +37,5 @@ out/
 .vscode/
 
 ### mail ###
-application-mail.yml
 application-mail.properties
+application-jwt.properties
