+++ src/main/java/com/youngchayoungcha/tastynote/service/EmailServiceImpl.java
@@ -1,0 +1,24 @@
+package com.youngchayoungcha.tastynote.service;
+
+import com.youngchayoungcha.tastynote.web.dto.MailCertifiedDTO;
+import lombok.RequiredArgsConstructor;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class EmailServiceImpl implements EmailService {
+    private final JavaMailSender javaMailSender;
+    private static final String FROM_ADDRESS = "jhcheon2134@gmail.com";
+
+    public void sendEmail(MailCertifiedDTO mailDTO) {
+        SimpleMailMessage message = new SimpleMailMessage();
+        message.setTo(mailDTO.getAddress());
+        message.setFrom(EmailServiceImpl.FROM_ADDRESS);
+        message.setSubject(mailDTO.getTitle());
+        message.setText("인증키 : " + mailDTO.getCertifiedKey());
+
+        javaMailSender.send(message);
+    }
+}
